+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [4.9.12] - 2021-03-18 13:09:00
+### Fixed
+- Customer-data loading
+- Upgrade Schema
+### Disabled
+- Staging updates support for Community edition
+
 ## [4.9.09] - 2021-02-08 13:09:00
 ### Added
 - Alternative stock field support